Inverted the plugin and factory class hierarchy for each sample plugin
